fix deepspeep can't find localhost
