+++ dataset/__init__.py
@@ -1,10 +1,11 @@
-from .deepseek_vl_sft_dataset import (
+from .deepseek_vl import (
     DeepSeekDPODataset,
     DeepSeekSftDataset,
     make_dpo_data_modlue,
     make_sft_data_modlue,
 )
 from .split import train_test_split
+
 
 __all__ = [
     "train_test_split",
