+++ scripts/train_deepseek_vl_dpo.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-gpus=1
+gpus=4,5
 
 project_name=deepseek_vl_7b_dpo
 model=deepseek_vl
@@ -8,8 +8,10 @@ dataset=deepseek_vl_dpo
 training=deepseek_vl_dpo
 
 
-WANDB_PROJECT=${project_name} CUDA_VISIBLE_DEVICES=${gpus} python deepseek_vl_dpo.py \
+WANDB_PROJECT=${project_name} deepspeed --include localhost:${gpus} deepseek_vl_dpo.py \
     project=${project_name} \
     model=${model} \
     dataset=${dataset} \
-    training=${training}
+    training=${training} \
+    model.lora.lora_enable=True \
+    training.deepspeed="scripts/deepspeed/zero3.json" \
