+++ utils/image.py
@@ -3,6 +3,6 @@ from io import BytesIO
 from PIL import Image
 
 
-def image_bytes2PIL(image_bytes):
+def image_bytes2PIL(image_bytes) -> Image.Image:
     image = Image.open(BytesIO(image_bytes))
     return image
